backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
ultralytics
torch
torchvision
roboflow